It turns out the big move to the role defaults location buried the certbot/letsencrypt-specific variables too deep.
In hindsight, this is obvious, as the proxy role is separate
and distinct from the certbot role - variables scoped within
one are not available to the other, as would be expected.

The fix for getting certbot variables into a place to be easily
overriden by the inventory file, yet still offering sane defaults
is to have them defined at the group_vars level.

Since this keeps rearing it's head, adding another reference to the
docs:
https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html#variable-precedence-where-should-i-put-a-variable